expand ParentID to int32 to prevent overflow in cluster-scale trees
int16 has a max of 32767. Cluster-scale halos (log M > 14) can produce
>32k branches, causing an OverflowError when storing branch IDs.

+++ TreeGen.py
@@ -81,7 +81,7 @@ def loop(itree):
     
     mass = np.zeros((cfg.Nmax,cfg.Nz)) - 99.
     order = np.zeros((cfg.Nmax,cfg.Nz),np.int8) - 99
-    ParentID = np.zeros((cfg.Nmax,cfg.Nz),np.int16) - 99
+    ParentID = np.zeros((cfg.Nmax,cfg.Nz),np.int32) - 99
     
     VirialRadius = np.zeros((cfg.Nmax,cfg.Nz),np.float32) - 99.
     concentration = np.zeros((cfg.Nmax,cfg.Nz),np.float32) - 99.
